Add automtic marking for when a submission is made
The 'test' field of the Assignment model has been changed form a text field to a file field for a marking script to be uploaded.

The marking server is hardcoded to localhost:4000.  Student and teacher solutions will be run on the marking server.  Results from these will be compared using a marking script, currently on the decentmark server.

:#